Move email selector into independent component
